refactor: hexagonal hygiene (pairing DI via Pool + relocate AuditSink port)
Drain 2 findings hexagonal de l'audit codebase 2026-05-18:

- 3.1 (impact fort) Extract get_pairing_manager() runtime deferred-import
  from core/stores/pairing_protocol.py. PairingManager now attached to Pool
  at bootstrap; commands/pairing/handlers.py reads pool.pairing_manager.
  Removes the last true core->infrastructure runtime exemption from
  .importlinter (slug: DEBT:importlinter-adr048-transition).

- 2.4 (impact faible) Move AuditSink Protocol from core/cli/ to core/ports/
  for consistency with LlmProvider, TtsProtocol, STTProtocol (capability
  port - consumes audit infrastructure via JetStreamAuditSink).
