+++ src/lyra/core/ports/__init__.py
@@ -17,11 +17,13 @@ Constraints
   alongside the sub-domain they serve.
 """
 
+from lyra.core.ports.audit_sink import AuditSink
 from lyra.core.ports.llm import LlmProvider, LlmResult
 from lyra.core.ports.stt import STTProtocol
 from lyra.core.ports.tts import TtsProtocol
 
 __all__ = [
+    "AuditSink",
     "LlmProvider",
     "LlmResult",
     "STTProtocol",
