+++ src/lyra/core/ports/audit_sink.py
