+++ src/lyra/core/stores/pairing_protocol.py
@@ -6,10 +6,8 @@ the dependency-inversion principle (ADR-059).
 
 ``PairingError`` is re-exported from ``pairing_config`` (pure core module).
 ``PairingManagerProtocol`` is the structural interface that command handlers
-depend on.
-``get_pairing_manager`` is a thin facade that defers the import of the
-infrastructure singleton so that ``commands/`` never has a direct
-``lyra.infrastructure`` import.
+depend on. Injection happens via ``Pool.pairing_manager`` at bootstrap
+(composition root) — no deferred import required (ADR-059 V4).
 """
 
 from __future__ import annotations
@@ -21,7 +19,6 @@ from lyra.core.stores.pairing_config import PairingConfig, PairingError
 __all__ = [
     "PairingError",
     "PairingManagerProtocol",
-    "get_pairing_manager",
 ]
 
 
@@ -40,16 +37,3 @@ class PairingManagerProtocol(Protocol):
     def check_rate_limit(self, identity_key: str) -> bool: ...
 
     def record_failed_attempt(self, identity_key: str) -> None: ...
-
-
-def get_pairing_manager() -> PairingManagerProtocol | None:
-    """Return the module-level PairingManager from infrastructure.
-
-    The import is deferred so that ``lyra.commands`` never takes a
-    compile-time dependency on ``lyra.infrastructure``.
-    """
-    from lyra.infrastructure.stores.pairing import (  # noqa: PLC0415 — DEBT:plc0415-deferred-import
-        get_pairing_manager as _get,
-    )
-
-    return _get()
